Change Integer to Long in repository

+++ src/main/java/ua/com/javarush/gnew/contactm/repository/ContactBookRepository.java
@@ -3,5 +3,5 @@ package ua.com.javarush.gnew.contactm.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import ua.com.javarush.gnew.contactm.entity.ContactBook;
 
-public interface ContactBookRepository extends JpaRepository<ContactBook, Integer> {
+public interface ContactBookRepository extends JpaRepository<ContactBook, Long> {
 }
